+++ src/has.c
@@ -47,14 +47,14 @@ SEXP FANSI_has(SEXP x, SEXP ctl, SEXP warn) {
   for(R_xlen_t i = 0; i < len; ++i) {
     FANSI_interrupt(i);
     SEXP chrsxp = STRING_ELT(x, i);
-    FANSI_check_enc(chrsxp, i);
+    FANSI_check_chrsxp(chrsxp, i);
     int res_tmp = FANSI_has_int(chrsxp, ctl_int);
     // no great, but need to watch out for NA_LOGICAL == INT_MIN
     if(res_tmp == -1 && warn_int) {
       res_tmp = -res_tmp;
       warning(
-        "Encountered invalid ESC sequence at index [%.0f], %s%s",
-        (double) i + 1,
+        "Encountered invalid ESC sequence at index [%jd], %s%s",
+        FANSI_ind(i),
         "see `?unhandled_ctl`; you can use `warn=FALSE` to turn ",
         "off these warnings."
       );