+++ tests/unitizer/tohtml.R
@@ -101,7 +101,7 @@ unitizer_sect("simple html conversion", {
 })
 unitizer_sect("Bright Colors", {
   sgr_to_html("hello\033[94;101m world\033[39m yow\033[49mza")
-  # oob color
+  # oob color (!98 %in% 90:97)
   sgr_to_html("hello\033[98;101m world\033[39m yow\033[49mza")
 })
 unitizer_sect("Corner cases", {
@@ -109,16 +109,33 @@ unitizer_sect("Corner cases", {
   sgr_to_html("hello\033[31m")
 
   # A string that shrinks; multiple repeated SGRs reduced to a single span
-
   sgrs <- paste0(rep("\033[31m", 20), collapse="")
   sgr_to_html(sprintf("%shello world\033[m", sgrs))
 
   # non character inputs
-
   sgr_to_html(1:3)
 
+  # Sequential escape sequences
+  sgr_to_html("\033[31mhello\033[m\033[42m world\033[m")
 
+  # Sequences in various spots
 
+  sgr_to_html("\033[33mhello")
+  sgr_to_html("he\033[33mllo")
+  sgr_to_html("hello\033[33m")
+
+  sgr_to_html(c("\033[33mhello", "world"))
+  sgr_to_html(c("\033[33mhello", "\033[44mworld"))
+  sgr_to_html(c("\033[33mhello", "wor\033[44mld"))
+  sgr_to_html(c("\033[33mhello", "world\033[44m"))
+  sgr_to_html(c("he\033[33mllo", "world"))
+  sgr_to_html(c("he\033[33mllo", "\033[44mworld"))
+  sgr_to_html(c("he\033[33mllo", "wor\033[44mld"))
+  sgr_to_html(c("he\033[33mllo", "world\033[44m"))
+  sgr_to_html(c("hello\033[33m", "world"))
+  sgr_to_html(c("hello\033[33m", "\033[44mworld"))
+  sgr_to_html(c("hello\033[33m", "wor\033[44mld"))
+  sgr_to_html(c("hello\033[33m", "world\033[44m"))
 })
 unitizer_sect("Bad inputs", {
   fansi:::esc_color_code_to_html(matrix(1:12, 4))
@@ -137,5 +154,53 @@ unitizer_sect("issue54", {
 
   string2 <- c("\033[3mhello\033[24m", "world\033[23m moon")
   fansi::sgr_to_html(string2)
+})
+unitizer_sect("Colors as classes (#65)", {
+  sgr_to_html("\033[94mhello\033[31;42;1mworld\033[m", classes=TRUE)
 
+  class.8 <-
+    do.call(paste, c(expand.grid(c("fg", "bg"), 0:7), sep="-"))
+  class.16 <-
+    do.call(paste, c(expand.grid(c("fg", "bg"), 0:15), sep="-"))
+  class.256 <-
+    do.call(paste, c(expand.grid(c("fg", "bg"), 0:255), sep="-"))
+
+  x <- c(
+    "\033[94mhe\033[107mllo\033[31;42;1mworld\033[m",
+    "\033[48;5;11;38;5;70mgood\033[7mbye\033[39;49m super \033[48;2;235;0;20mmoon\033[m",
+    NULL
+  )
+  term.cap <- c('bright', '256', 'truecolor')
+  sgr_to_html(x, classes=class.8, term.cap=term.cap)
+  sgr_to_html(x, classes=class.16, term.cap=term.cap)
+  sgr_to_html(x, classes=class.256, term.cap=term.cap)
+
+  make_styles(class.8)
+  make_styles(class.8, matrix(c(0,1,0,0,0,1,1,0,0), 3)) # shift channels
+
+  # in_html(sgr_to_html(sgr_256()))
+  sgr_to_html(sgr_256())
+  # in_html(sgr_to_html(sgr_256(), classes=make_styles(class.256)))
+  sgr_to_html(sgr_256(), classes=class.256)
+
+  # errors
+  sgr_to_html("\033[31mhello\033[31m", classes=NULL)
+  sgr_to_html("\033[31mhello\033[31m", classes=character(7L))
+  sgr_to_html("\033[31mhello\033[31m", classes=rep(NA_character_, 16))
+  sgr_to_html("\033[31mhello\033[31m", classes=rep("bad class", 16))
+
+  make_styles(class.8, c(1,1,0,0,0,1,1,0,0))
+  make_styles(class.8, matrix(c(0,1,0,0,0,1,1,0,NA), 3))
+  make_styles(class.8, "hello")
+  make_styles(letters, matrix(c(0,1,0,0,0,1,1,0,0), 3))
+  make_styles(NULL)
+
+  ## see examples for visual testing
+})
+unitizer_sect("helpers", {
+  html <- sgr_to_html("\033[42mHello")
+  f <- in_html(html, css="span {background-color: #CCC;}", display=FALSE)
+  readLines(f)
+  unlink(f)
+  in_html(html, css="span {background-color: #CCC;}", display=FALSE, clean=TRUE)
 })